refactor: merge load_cache and invalidation into reactive/cache.py
