+++ pyjinhx/__init__.py
@@ -42,9 +42,8 @@ from pyjinhx.reactive.dev import (
     enable_reactive_dev,
     format_dependency_graph,
 )
-from pyjinhx.reactive.invalidation import InvalidationBackend, InvalidationHub
+from pyjinhx.reactive.cache import CacheScope, InvalidationBackend, InvalidationHub, LoadCache
 from pyjinhx.reactive.keys import StateKey
-from pyjinhx.reactive.load_cache import CacheScope, LoadCache
 from pyjinhx.reactive.mutations import MutationTracker, mutates
 from pyjinhx.reactive.oob import oob_swaps
 from pyjinhx.reactive.pjx_load import PjxLoad
