+++ pyjinhx/config/__init__.py
@@ -7,9 +7,8 @@ from contextlib import contextmanager
 from dataclasses import dataclass, fields, replace
 from typing import Any
 
-from pyjinhx.reactive.load_cache import CacheScope, LoadCache
+from pyjinhx.reactive.cache import CacheScope, LoadCache, InvalidationBackend, InvalidationHub
 from pyjinhx.reactive.dev import disable_reactive_dev, enable_reactive_dev
-from pyjinhx.reactive.invalidation import InvalidationBackend, InvalidationHub
 
 
 logger = logging.getLogger("pyjinhx")
