+++ pyjinhx/reactive/oob.py
@@ -11,7 +11,7 @@ from pyjinhx.reactive.keys import ReactiveKey, coerce_reactive_keys
 from pyjinhx.utils import css_attribute_selector_attr_value, stamp_root_attributes
 
 from .client import MountedManifest
-from .load_cache import LoadCache
+from .cache import LoadCache
 
 
 @dataclass
