+++ pyjinhx/reactive/render.py
@@ -6,7 +6,7 @@ from markupsafe import Markup
 
 from .client import ClientBackend, MountedManifest, TriggerManifest
 from .dev import warn_reactive_render_without_client
-from .load_cache import LoadCache
+from .cache import LoadCache
 from .mutations import MutationTracker
 from .oob import oob_swaps
